fix html accounting table in table bug

+++ tests/tdnet/test_api.py
@@ -29,19 +29,20 @@ class TestAPI(unittest.TestCase):
         documents = xbrr.tdnet.api.documents.get(str(datetime.date.today()))
         self.assertEqual(documents.metadata.count, len(documents.list))
 
-        _dir = os.path.dirname(__file__)
-        d = documents.list[0]
-        for ext in ["xbrl", "pdf"]:
-            time.sleep(3)
-            if ext == "xbrl":
-                if not d.has_xbrl: continue
-                path = d.get_xbrl(save_dir=_dir)
-            else:
-                if not d.has_pdf: continue
-                path = d.get_pdf(save_dir=_dir)
+        if len(documents.list) > 0:
+            _dir = os.path.dirname(__file__)
+            d = documents.list[0]
+            for ext in ["xbrl", "pdf"]:
+                time.sleep(3)
+                if ext == "xbrl":
+                    if not d.has_xbrl: continue
+                    path = d.get_xbrl(save_dir=_dir)
+                else:
+                    if not d.has_pdf: continue
+                    path = d.get_pdf(save_dir=_dir)
 
-            self.assertTrue(os.path.exists(path))
-            os.remove(path)
+                self.assertTrue(os.path.exists(path))
+                os.remove(path)
 
     def test_extract(self):
         path = os.path.join(os.path.dirname(__file__),
